Fixed validation errors from hidden input fields in edit table dialog being validated when they should be ignored.
