+++ PWA/src/components/Timesheet/IDDForm.vue
@@ -132,6 +132,7 @@
         <v-col>
           <ConfirmSubmission
             :valid="valid"
+            :errors="errors"
             :formFields="formFields"
             :totalEdited="totalEdited"
             :validationSignal="validationSignal"
@@ -149,6 +150,7 @@
   import ConfirmSubmission from "@/components/Timesheet/ConfirmSubmission";
   import fieldData from "@/components/Timesheet/IDDFormFields.json";
   import rules from "@/components/Timesheet/FormRules.js";
+  import time_functions from "@/components/Timesheet/TimeFunctions.js";
 
   export default {
     name: "IDDForm",
@@ -206,6 +208,9 @@
         // Hide form validation error messages by default
         valid: true,
 
+        // The number of invalid fields
+        errors: [],
+
         // Signal denoting completion of validation for form fields
         validationSignal: false,
       };
@@ -250,10 +255,117 @@
 
       // Check if the form fields have valid input
       validateInputs() {
-        this.$refs.form.validate();
+        this.validate(); 
 
         // Change this field to validation signal ConfirmSubmission
         this.validationSignal = !this.validationSignal;
+      },
+
+      // Compute the sum of all serviceDeliveredOn totalHours with the totalHours field
+      sumTableHours() {
+        var sumHours = 0;
+        var sumMinutes = 0;
+
+        // For each row in the array of entries...
+        this.formFields["serviceDeliveredOn"]["value"].forEach((entry) => {
+          // Check that the totalHours field is valid
+          if (entry["errors"]["totalHours"].length == 0) {
+            sumHours += parseInt(entry["totalHours"].substr(0, 2));
+            sumMinutes += parseInt(entry["totalHours"].substr(3, 2));
+          }
+        });
+        sumHours += (sumMinutes - (sumMinutes % 60)) / 60;
+        sumMinutes %= 60;
+
+        return sumHours.toString() + ":" + sumMinutes.toString();
+      },
+
+      // Count the number of errors in the serviceDeliveredOn table
+      getTableErrors() {
+        // For each row in the array of entries...
+        this.formFields["serviceDeliveredOn"]["value"].forEach(
+          (entry, index) => {
+            // For each error col in an entry, check the amount of errors
+            Object.entries(entry["errors"]).forEach(([col, errors]) => {
+              var colErrors = errors.length;
+              if (colErrors > 0) {
+                this.errors.push(
+                  [`ERROR: in row ${
+                    index + 1
+                  } of the serviceDeliveredOn table, '${col}' has the following errors:`,
+                  errors]
+                );
+              }
+            });
+          }
+        );
+      },
+
+      // Validate the form
+      validate() {
+        // Reset all error messages and validation
+        this.errors = [];
+
+        // Check parent's response on validity of input fields
+        if (!this.$refs.form.validate()) {
+          this.errors.push("ERROR: Invalid input in some form fields!");
+        }
+
+
+        // Check the validity of the serviceDeliveredOn table
+        this.getTableErrors();
+
+        // Ensure that the serviceDeliveredOn table sum == totalHours field
+        if (this.formFields.totalHours.value !== null) {
+          var sumHours = this.sumTableHours();
+          if (sumHours.localeCompare(this.formFields.totalHours.value) !== 0) {
+            this.errors.push(
+              `ERROR: valid rows in the serviceDeliveredOn table sums up to ${sumHours} hours, but the totalHours field reports ${this.formFields.totalHours.value} hours!`
+            );
+          }
+        }
+
+        // If there were no edited fields, ensure that the provider and
+        // employer signature date are after the last service date
+        if (this.totalEdited <= 0) {
+          // Only compare the earlier date
+          var comparisonDate = this.formFields.providerSignDate.value;
+          if (
+            time_functions.dateCompare(
+              comparisonDate,
+              this.formFields.employerSignDate.value
+            ) > 0
+          ) {
+            comparisonDate = this.formFields.employerSignDate.value;
+          }
+
+          // Compare signage dates with the pay period
+          // Note, only comparing the YYYY-mm part
+          var submissionDate = this.formFields.submissionDate.value;
+          if (
+            time_functions.dateCompare(
+              comparisonDate.substr(0, 7),
+              submissionDate
+            ) < 0
+          ) {
+            this.errors.push(
+              `ERROR: the employer or provider sign date is before the pay period.`
+            );
+          }
+
+          // Get the last date from the serviceDeliveredOn table
+          var latestDateIdx = this.formFields.serviceDeliveredOn.value.length;
+          if (latestDateIdx > 0) {
+            var latestDate = this.formFields.serviceDeliveredOn.value[
+              latestDateIdx - 1
+            ]["date"];
+            if (time_functions.dateCompare(comparisonDate, latestDate) < 0) {
+              this.errors.push(
+                `ERROR: the employer or provider sign date is before the latest service delivery date.`
+              );
+            }
+          }
+        }
       },
 
       reset() {
